requirements.txt
Django==1.9.7
django-cors-headers==1.1.0
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
drf-extensions==0.2.8
mysqlclient==1.3.7
django-filter==0.13.0
Pillow==3.2.0
pytz==2016.4
